pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "koa"
version = "0.5.0"
description = "CLI for submitting and managing KOA HPC jobs from a local machine."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
  {name = "Michael Buriek"},
  {name = "Yosub Shin"}
]
dependencies = [
  "PyYAML>=6.0.1",
  "rich>=13.0.0",
  "streamlit>=1.32.0",
  "streamlit-aggrid>=0.3.4.post3",
  "pandas>=2.0.0",
  "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "black",
  "ruff",
  "pytest>=7.0",
  "pytest-mock>=3.10",
]

[project.scripts]
koa = "koa.__main__:main"
koa-mcp = "koa.mcp_server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
koa = ["templates/*.tmpl"]
